+++ shiki_organizer/cli/__main__.py
@@ -1,13 +1,12 @@
 import datetime as dt
-import math
+import string
 import uuid
 
 import click
 from colorama import Fore, Style
-from termcolor import colored
 
 from shiki_organizer.actions import get_status
-from shiki_organizer.model import Interval, Task
+from shiki_organizer.model import Interval, Project, Tag, Task
 
 
 @click.group()
@@ -16,22 +15,16 @@ def cli():
 
 
 @click.command()
-@click.argument("name")
+@click.argument("description")
 @click.option(
-    "-p", "--priority", default=0, help="Priority of the task.", type=click.IntRange(0)
-)
-@click.option(
-    "-i",
-    "--divider",
-    default=1,
-    help="Divider of the task score.",
-    type=click.IntRange(0),
+    "-p",
+    "--priority",
+    help="Priority of the task.",
+    type=click.Choice(list(string.ascii_uppercase)),
 )
 @click.option(
     "-r",
     "--recurrence",
-    is_flag=False,
-    flag_value=1,
     help="Recurrence interval of the task.",
     type=click.IntRange(1),
 )
@@ -52,26 +45,29 @@ def cli():
     type=click.DateTime(formats=["%Y-%m-%d"]),
 )
 @click.option(
-    "-a",
-    "--parent",
-    help="Id or uuid of the parent task.",
+    "-t",
+    "--tags",
+    multiple=True,
+    help="Tags of the task.",
     type=str,
 )
-def add(name, priority, divider, recurrence, scheduled, deadline, parent):
-    if parent:
-        if parent.isnumeric():
-            parent = Task.get_by_id(int(parent))
-        else:
-            parent = uuid.UUID(parent)
+@click.option(
+    "--project",
+    help="Project of the task.",
+    type=str,
+)
+def add(description, priority, recurrence, scheduled, deadline, tags, project):
+    project, _ = Project.get_or_create(name=project) if project else (None, None)
+    tags = [tag for tag, _ in [Tag.get_or_create(name=tag) for tag in tags]]
     task = Task.create(
-        name=name,
+        description=description,
         priority=priority,
-        divider=divider,
         recurrence=recurrence,
         scheduled=scheduled.date() if scheduled else None,
         deadline=deadline.date() if deadline else None,
-        parent=parent,
+        project=project,
     )
+    task.tags.add(tags)
     Task.reindex()
     task = Task.get_by_uuid(task.uuid)
     print(f"Created task {task.id}.")
@@ -79,23 +75,19 @@ def add(name, priority, divider, recurrence, scheduled, deadline, parent):
 
 @click.command()
 @click.argument("task")
-@click.option("-n", "--name", help="Name of a task.")
+@click.option(
+    "-n",
+    "--description",
+    help="Description of the task.",
+    type=str,
+)
 @click.option(
     "-p",
     "--priority",
     is_flag=False,
-    flag_value=0,
+    flag_value="",
     help="Priority of the task.",
-    type=click.IntRange(0),
-)
-@click.option(
-    "-i",
-    "--divider",
-    is_flag=False,
-    flag_value=1,
-    default=1,
-    help="Divider of the task score.",
-    type=click.IntRange(0),
+    type=click.Choice(list(string.ascii_uppercase) + [""]),
 )
 @click.option(
     "-r",
@@ -103,7 +95,7 @@ def add(name, priority, divider, recurrence, scheduled, deadline, parent):
     is_flag=False,
     flag_value=0,
     help="Recurrence interval of the task.",
-    type=click.IntRange(0),
+    type=int,
 )
 @click.option(
     "-s",
@@ -122,64 +114,92 @@ def add(name, priority, divider, recurrence, scheduled, deadline, parent):
     type=click.DateTime(formats=["%Y-%m-%d"]),
 )
 @click.option(
-    "-a",
-    "--parent",
+    "-t",
+    "--tags",
     is_flag=False,
     flag_value="",
-    help="Id or uuid of the parent task.",
+    multiple=True,
+    help="Tags of the task.",
     type=str,
 )
-def modify(task, name, priority, divider, recurrence, scheduled, deadline, parent):
+@click.option(
+    "--project",
+    is_flag=False,
+    flag_value="",
+    help="Project of the task.",
+    type=str,
+)
+def modify(task, description, priority, recurrence, scheduled, deadline, tags, project):
     if task.isnumeric():
         task = Task.get_by_id(int(task))
     else:
         task = Task.get_by_uuid(uuid.UUID(task))
-    name = task.name if name == None else name
-    priority = task.priority if priority == None else priority
-    divider = task.divider if divider == None else divider
-    if recurrence == 0:
-        recurrence = None
-    elif recurrence == None:
+    description = task.description if description == None else description
+    if priority == None:
+        priority = task.priority
+    elif priority == "":
+        priority = None
+    if recurrence == None:
         recurrence = task.recurrence
-    else:
-        recurrence
-    if scheduled == dt.date.min:
-        scheduled = None
-    elif scheduled == None:
+    elif recurrence == 0:
+        recurrence = None
+    if scheduled == None:
         scheduled = task.scheduled
+    elif scheduled == dt.datetime.min:
+        scheduled = None
     else:
-        scheduled.date()
-    if deadline == dt.date.min:
-        deadline = None
-    elif deadline == None:
+        scheduled = scheduled.date()
+    if deadline == None:
         deadline = task.deadline
+    elif deadline == dt.datetime.min:
+        deadline = None
     else:
-        deadline.date()
-    if parent == "":
-        parent = None
-    elif parent == None:
-        parent = task.parent
+        deadline = deadline.date()
+    if len(tags) == 0:
+        pass
+    elif len(tags) == 1 and tags[0] == "":
+        task.tags.clear()
     else:
-        if parent.isnumeric():
-            parent = Task.get_by_id(int(parent))
-        else:
-            parent = uuid.UUID(parent)
+        task.tags.clear()
+        tags = [tag for tag, _ in [Tag.get_or_create(name=tag) for tag in tags]]
+        task.tags.add(tags)
+    if project == None:
+        project = task.project
+    elif project == "":
+        project = None
+    else:
+        project, _ = Project.get_or_create(name=project)
     q = task.update(
-        name=name,
+        description=description,
         priority=priority,
-        divider=divider,
         recurrence=recurrence,
         scheduled=scheduled,
         deadline=deadline,
-        parent=parent,
+        project=project,
     ).where(Task.id == task.id)
     q.execute()
-    print(f"Modifying task {task.id} '{task.name}'.")
+    Project.remove_unused()
+    Tag.remove_unused()
+    print(f"Modifying task {task.id} '{task.description}'.")
 
 
 @click.command()
-@click.option("-t", "--task", type=str, help="Id or uuid of the task")
-@click.option("-d", "--description", type=str, help="Іnterval description")
+@click.argument("task")
+def delete(task):
+    if task.isnumeric():
+        task = Task.get_by_id(int(task))
+    else:
+        task = Task.get_by_uuid(uuid.UUID(task))
+    task.tags.clear()
+    Interval.delete().where(Interval.task == task).execute()
+    task.delete_instance()
+    Project.remove_unused()
+    Tag.remove_unused()
+    print(f"Deleting task {task.id} '{task.description}'.")
+
+
+@click.command()
+@click.argument("task")
 @click.option(
     "-s",
     "--start",
@@ -193,7 +213,7 @@ def modify(task, name, priority, divider, recurrence, scheduled, deadline, paren
     type=click.DateTime(formats=["%Y-%m-%dT%H:%M:%S"]),
     help="End of the interval",
 )
-def start(task, description, start, end):
+def start(task, start, end):
     if task.isnumeric():
         task = Task.get_by_id(int(task))
     else:
@@ -214,7 +234,7 @@ def start(task, description, start, end):
         end = None
         print("The end of the interval must be after the start.")
         return
-    Interval.create(task=task, start=start, end=end, description=description)
+    Interval.create(task=task, start=start, end=end)
     print(get_status())
     Interval.reindex()
 
@@ -224,148 +244,13 @@ def stop():
     interval = Interval.get_or_none(Interval.end == None)
     if interval:
         print(get_status("Recorded"))
-        interval.end = dt.datetime.now()
+        now = dt.datetime.now()
+        interval.end = dt.datetime(
+            now.year, now.month, now.day, now.hour, now.minute, now.second
+        )
         interval.save()
     else:
         print("There is no active time tracking.")
-
-
-@click.command()
-@click.option("-s", "--hide-scheduled", type=bool, help="ide tasks with scheduled")
-@click.option(
-    "-t/-n", "--today/--no-today", default=False, help="Show only today tasks"
-)
-def tree(today, hide_scheduled):
-    def row_to_str(key, row):
-        if row["task"].divider != 1:
-            divider = f'({row["task"].divider}) '
-        else:
-            divider = ""
-        if row["task"].priority != 0:
-            priority = f'({row["task"].priority}) '
-        else:
-            priority = ""
-        scheduled = row["task"].scheduled
-        scheduled = (
-            colored(f" scheduled:", "magenta") + str(scheduled) if scheduled else ""
-        )
-        deadline = row["task"].deadline
-        deadline = colored(f" deadline:", "red") + str(deadline) if deadline else ""
-        recurrence = row["task"].recurrence
-        recurrence = (
-            colored(f" recurrence:", "yellow") + str(recurrence) + "d"
-            if recurrence
-            else ""
-        )
-        duration = colored(f" duration:", "green") + f"{round(row['duration']/60)}"
-        days = colored(f" days:", "blue") + f"{row['days']}"
-        avg = colored(f" avg:", "cyan") + f"{round(row['avg']/60)}"
-        score = colored(f" score:", "yellow") + f"{round(row['score']/60)}"
-        return f'{divider}{priority}{key} {row["task"].name}{scheduled}{recurrence}{deadline}{duration}{days}{avg}{score}'
-
-    def sort_queue(queue):
-        if hide_scheduled:
-            queue = list(filter(lambda x: table[x[1]]["task"].scheduled == None, queue))
-        for x in queue:
-            task = table[x[1]]["task"]
-            if task.divider != 0:
-                table[x[1]]["score"] = table[x[1]]["duration"] / task.divider
-            else:
-                table[x[1]]["score"] = -1
-        if today:
-            queue = sorted(
-                queue,
-                key=lambda x: table[x[1]]["task"].priority
-                if table[x[1]]["task"].priority != 0
-                else float("infinity"),
-                reverse=True,
-            )
-        else:
-            queue = sorted(queue, key=lambda x: table[x[1]]["score"], reverse=True)
-            queue = list(filter(lambda x: table[x[1]]["score"] != -1, queue))
-        return queue
-
-    table = {}
-    for task in Task.select():
-        table[task.id] = {
-            "task": task,
-            "duration": 0,
-            "days": set(),
-            "avg": 0,
-            "score": 0,
-        }
-    for interval in Interval.select():
-        duration = interval.duration
-        date = str(interval.start.date())
-        for task in interval.task.parents + [interval.task]:
-            if task.id in table:
-                table[task.id]["days"].add(date)
-                table[task.id]["duration"] += duration
-    for key in table:
-        table[key]["days"] = len(table[key]["days"])
-        if table[key]["days"]:
-            table[key]["avg"] = table[key]["duration"] / table[key]["days"]
-        else:
-            table[key]["avg"] = 0
-    table = dict(sorted(table.items(), key=lambda tag: tag[1]["score"]))
-    if today:
-        table = dict(
-            sorted(
-                table.items(),
-                key=lambda tag: tag[1]["task"].scheduled
-                if tag[1]["task"].scheduled != None
-                else dt.date.max,
-            )
-        )
-        queue = [
-            (0, task_id)
-            for task_id in table
-            if table[task_id]["task"].scheduled == dt.date.today()
-        ]
-    else:
-        queue = [
-            (0, task_id) for task_id in table if table[task_id]["task"].parent == None
-        ]
-    queue = sort_queue(queue)
-    while queue:
-        row = queue.pop()
-        children = [(row[0] + 1, t.id) for t in table[row[1]]["task"].direct_children]
-        children = sort_queue(children)
-        queue += children
-        if table[row[1]]["task"].archived == False:
-            print(" " * 4 * row[0], row_to_str(row[1], table[row[1]]), sep="")
-
-
-@click.command()
-@click.argument("task")
-def delete(task):
-    if task.isnumeric():
-        task = Task.get_by_id(int(task))
-    else:
-        task = Task.get_by_uuid(uuid.UUID(task))
-    task.delete_instance()
-    print(f"Deleting task {task.id} '{task.name}'.")
-    Task.reindex()
-
-
-@click.command()
-def normalize():
-    parents = {None: 1}
-    count = 0
-
-    for parent in Task.select():
-        if Task.select().where(Task.parent == parent).count() > 2:
-            parents[parent] = 1
-    for parent in parents:
-        tasks = Task.select().where(Task.parent == parent)
-        divider = [task.divider for task in tasks]
-        gcd = math.gcd(*divider)
-        if gcd > 1:
-            for task in tasks:
-                task.divider /= gcd
-            Task.bulk_update(tasks, [Task.divider])
-            count += len(tasks)
-    print(f"Modified {count} tasks.")
 
 
 @click.command()
@@ -396,7 +281,7 @@ def done(task):
     else:
         task.archived = True
     task.save()
-    print(f"Completed task {task.id} '{task.name}'.")
+    print(f"Completed task {task.id} '{task.description}'.")
 
 
 @click.command()
@@ -409,21 +294,15 @@ def interval_list(uuid):
             uuid = ""
         id = f"{Fore.RED}id:{Style.RESET_ALL}{interval.id} "
         if interval.task:
-            task = f"{Fore.YELLOW}task:{Style.RESET_ALL}'{interval.task.name}' "
+            task = f"{Fore.YELLOW}task:{Style.RESET_ALL}'{interval.task.description}' "
         else:
             task = ""
-        if interval.description:
-            description = (
-                f"{Fore.GREEN}description:{Style.RESET_ALL}{interval.description} "
-            )
-        else:
-            description = ""
         start = f"{Fore.BLUE}start:{Style.RESET_ALL}'{interval.start}' "
         if interval.end:
             end = f"{Fore.MAGENTA}end:{Style.RESET_ALL}'{interval.end}'"
         else:
             end = ""
-        print(f"{uuid}{id}{task}{description}{start}{end}")
+        print(f"{uuid}{id}{task}{start}{end}")
 
 
 @click.command()
@@ -434,14 +313,6 @@ def interval_list(uuid):
     is_flag=False,
     flag_value="",
     help="Id or uuid of the parent task.",
-    type=str,
-)
-@click.option(
-    "-d",
-    "--description",
-    is_flag=False,
-    flag_value="",
-    help="Description of interval.",
     type=str,
 )
 @click.option(
@@ -458,7 +329,7 @@ def interval_list(uuid):
     type=click.DateTime(formats=["%Y-%m-%dT%H:%M:%S"]),
     help="End of the interval",
 )
-def modify_interval(interval, task, description, start, end):
+def interval_modify(interval, task, start, end):
     if interval.isnumeric():
         interval = Interval.get_by_id(int(interval))
     else:
@@ -472,10 +343,6 @@ def modify_interval(interval, task, description, start, end):
             task = Task.get_by_id(int(task))
         else:
             task = Task.get_by_uuid(uuid.UUID(task))
-    if description == None:
-        description = interval.description
-    elif description == "":
-        description = None
     if start == None:
         start = interval.start
     if end == None:
@@ -484,7 +351,6 @@ def modify_interval(interval, task, description, start, end):
         end = None
     q = interval.update(
         task=task,
-        description=description,
         start=start,
         end=end,
     ).where(Interval.id == interval.id)
@@ -494,7 +360,7 @@ def modify_interval(interval, task, description, start, end):
 
 @click.command()
 @click.argument("interval")
-def delete_interval(interval):
+def interval_delete(interval):
     if interval.isnumeric():
         interval = Interval.get_by_id(int(interval))
     else:
@@ -504,18 +370,145 @@ def delete_interval(interval):
     Interval.reindex()
 
 
+@click.command(name="ls")
+@click.option(
+    "-t", "--today", is_flag=True, default=False, help="Show only today tasks."
+)
+@click.option(
+    "-p",
+    "--project",
+    is_flag=False,
+    flag_value="",
+    type=str,
+    help="Filter tasks by project.",
+)
+@click.option(
+    "--tag", type=str, is_flag=False, flag_value="", help="Filter tasks by tag."
+)
+def ls(today, project, tag):
+    def task_to_str(task):
+        if task.priority:
+            priority = f"{Fore.BLUE}({task.priority}){Style.RESET_ALL} "
+        else:
+            priority = ""
+        scheduled = (
+            f" {Fore.MAGENTA}scheduled:{Style.RESET_ALL}" + str(task.scheduled)
+            if task.scheduled
+            else ""
+        )
+        deadline = (
+            f" {Fore.RED}deadline:{Style.RESET_ALL}" + str(task.deadline)
+            if task.deadline
+            else ""
+        )
+        recurrence = (
+            f" {Fore.YELLOW}recurrence:{Style.RESET_ALL}{task.recurrence}d"
+            if task.recurrence
+            else ""
+        )
+        project = (
+            f" {Fore.CYAN}project:{Style.RESET_ALL}{task.project.name}"
+            if task.project
+            else ""
+        )
+        tags = (
+            " ".join([f"{Fore.GREEN}+{tag.name}{Style.RESET_ALL}" for tag in task.tags])
+            if task.tags
+            else ""
+        )
+        return f"{task.id} {priority}{task.description}{scheduled}{recurrence}{deadline}{project} {tags}"
+
+    tasks = None
+    if project:
+        tasks = Task.select().join(Project).where(Project.name == project)
+    if project == "":
+        tasks = Task.select().where(Task.project == None)
+    if tag != None:
+        tag = None if not tag else tag
+        tasks = []
+        for task in Task.select():
+            tag_nams = [tag.name for tag in task.tags]
+            if tag:
+                if tag in tag_nams:
+                    tasks.append(task)
+            elif len(tag_nams) == 0:
+                tasks.append(task)
+    if not tasks:
+        tasks = Task.select()
+    if today:
+        tasks = tasks.where(Task.scheduled == dt.date.today())
+    tasks = sorted(tasks, key=lambda x: x.priority if x.priority else "a")
+    tasks = sorted(tasks, key=lambda x: x.scheduled if x.scheduled else dt.date.max)
+    for task in tasks:
+        print(task_to_str(task))
+
+
+@click.command()
+def tags():
+    tags = {}
+    global_duration = 0
+    for interval in Interval.select():
+        for tag in interval.task.tags:
+            duration = interval.duration
+            if tag in tags:
+                tags[tag] += duration
+            else:
+                tags[tag] = duration
+            global_duration += duration
+    tags = sorted(tags.items(), key=lambda x: x[1], reverse=True)
+    for tag, duration in tags:
+        print(
+            tag.name,
+            round(duration / 60 * 100) / 100,
+            f"{round(duration / global_duration * 10000)/100}%",
+        )
+
+
+@click.command()
+def projects():
+    projects = {}
+    global_duration = 0
+    for interval in Interval.select():
+        duration = interval.duration
+        project = interval.task.project
+        if project:
+            if project in projects:
+                projects[project] += duration
+            else:
+                projects[project] = duration
+            global_duration += duration
+    projects = sorted(projects.items(), key=lambda x: x[1], reverse=True)
+    for project, duration in projects:
+        print(
+            project.name,
+            round(duration / 60 * 100) / 100,
+            f"{round(duration / global_duration * 10000)/100}%",
+        )
+
+
+@click.command()
+def today():
+    global_duration = 0
+    for interval in Interval.select():
+        if interval.start.date() == dt.date.today():
+            global_duration += interval.duration
+    print(f"{round(global_duration / 60 / 60 * 100) / 100}h")
+
+
 cli.add_command(add)
 cli.add_command(modify)
+cli.add_command(delete)
 cli.add_command(start)
 cli.add_command(stop)
-cli.add_command(tree)
-cli.add_command(delete)
-cli.add_command(normalize)
 cli.add_command(status)
 cli.add_command(done)
 cli.add_command(interval_list)
-cli.add_command(modify_interval)
-cli.add_command(delete_interval)
+cli.add_command(interval_modify)
+cli.add_command(interval_delete)
+cli.add_command(ls)
+cli.add_command(tags)
+cli.add_command(projects)
+cli.add_command(today)
 
 
 def main():